refactored locations of controls, updated formatting
