+++ src/App.js
@@ -2,6 +2,7 @@ import "./App.css";
 import { Spells } from "./CharacterSheetItems/Spells/Spell";
 import { CharacterStats } from "./CharacterSheetItems/CharacterStats/CharacterStats.js";
 import CharacterSheetData from "./CharacterSheet.json";
+import { GreyBackGroundBorder } from "./Controls/GreyBackGroundBorder.js";
 function App() {
   //https://github.com/cynicaloptimist/improved-initiative Gh copilot pulled from here??
   // Mit License
@@ -11,15 +12,17 @@ function App() {
       {CharacterSheetData.spells.map((spellset) => (
         <Spells spells={spellset} key={spellset.name}></Spells>
       ))}
-      {Object.keys(CharacterSheetData.Stats).map((key) => (
-        <CharacterStats
-          key={key}
-          title={key}
-          value={CharacterSheetData.Stats[key].value}
-          modifier={CharacterSheetData.Stats[key].modifier}
-          skills={CharacterSheetData.Stats[key].skills}
-        ></CharacterStats>
-      ))}
+      <GreyBackGroundBorder>
+        {Object.keys(CharacterSheetData.Stats).map((key) => (
+          <CharacterStats
+            key={key}
+            title={key}
+            value={CharacterSheetData.Stats[key].value}
+            modifier={CharacterSheetData.Stats[key].modifier}
+            skills={CharacterSheetData.Stats[key].skills}
+          ></CharacterStats>
+        ))}
+      </GreyBackGroundBorder>
     </div>
   );
 }
