+++ src/Controls/GreyBackGroundBorder.js
@@ -1,0 +1,6 @@
+import React from "react";
+import "./GreyBackGroundborder.css";
+
+export function GreyBackGroundBorder(props) {
+  return <div className="grey-back-ground-border">{props.children}</div>;
+}
